Creacion del menu visible para el usuario, añadido a la base de datos un unico administrador, añadido la posibilidad al administrador para utilizar el metodo crud, clase usuario y administrador han sido modificadas
